requirements.txt
# SQL Interview Prep App - Python Dependencies

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Environment variable management
python-dotenv==1.0.0
